requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
commonmark==0.9.1
docutils==0.20.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
rich==12.6.0
rich-rst==1.1.7
textual==0.1.18
tqdm==4.66.1
urllib3==2.0.5
yarl==1.9.2
